requirements.txt
streamlit
langchain
langchain-community
langchain-core
langchain-groq
google-api-python-client
pydantic
python-dotenv